reoganized directory structure, updated Makefile.

+++ Makefile
@@ -1,5 +1,5 @@
 
-test:
+test: words/words.go filelist/filelist.go cmds/makewordlist/makewordlist.go
 	gofmt -w words/words.go
 	golint words/words.go
 	gofmt -w words/words_test.go
@@ -10,14 +10,17 @@ test:
 	gofmt -w filelist/filelist_test.go
 	golint filelist/filelist_test.go
 	go test -v ./filelist
-	gofmt -w makewordlist.go
-	golint makewordlist.go
+	gofmt -w cmds/makewordlist/makewordlist.go
+	golint cmds/makewordlist/makewordlist.go
 
-build: makewordlist.go words/words.go filelist/filelist.go
-	go build makewordlist.go
+build: cmds/makewordlist/makewordlist.go words/words.go filelist/filelist.go
+	go build cmds/makewordlist/makewordlist.go
 
-clean: makewordlist
-	rm makewordlist
+clean:
+	if [ -f makewordlist ]; then rm makewordlist; fi
+	if [ -f files.json ]; then rm files.json; fi
+	if [ -f wordlist.json ]; then rm wordlist.json; fi
 
-install: makewordlist.go
-	go install makewordlist.go
+install: cmds/makewordlist/makewordlist.go
+	go install cmds/makewordlist/makewordlist.go
+
